% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/closestDefi.R
\name{closestDefi}
\alias{closestDefi}
\title{Return the coordinates and the distance to the next defibrillator}
\usage{
closestDefi(cityname, x, y)
}
\arguments{
\item{cityname}{a character value - name of the city.}

\item{x}{A numeric value - x coordinate of the current position.}

\item{y}{A numeric value - y coordinate of the current position.}
}
\value{
A sentence - coordinates and the distance to the next difibrillator.
}
\description{
This function takes two numeric values and a character value as input and returns a text.
}
\examples{
\dontrun{
closestDefi("München",11.50,48.10)
# Output: The distance to the next defibrillator is located in 529 meters.The coordinates are: (11.4976269,48.0957883).
closestDefi("München",11.60,48.20)
# Output: The distance to the next defibrillator is located in 1384 meters.The coordinates are: (11.609247,48.1915003).
closestDefi("München",11.70,48.07)
# The distance to the next defibrillator is located in 2026 meters.The coordinates are: (11.6825927,48.0645139).
}
}
