% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDefi.R
\name{plotDefi}
\alias{plotDefi}
\title{Plotting all the defibrillators in the city}
\usage{
plotDefi(cityname)
}
\arguments{
\item{cityname}{a character value - name of the city.}
}
\value{
a tmap plot.
}
\description{
This function takes one character value as input and returns a tmap-plot.
}
\examples{
\dontrun{
plotDefi("München")
# Output: plot of defibrillators in Munich
plotDefi("Berlin")
# Output: plot of defibrillators in Berlin
}
}
